refact: separate single file into several other files using namespaces
