+++ src/components/project-input.component.ts
@@ -1,0 +1,91 @@
+/// <reference path='base.component.ts' />
+/// <reference path='../decorators/autobind.decorator.ts' />
+/// <reference path='../util/validation.util.ts' />
+/// <reference path='../state/project.state.ts' />
+
+namespace App {
+    const titleValidation: Validation = {
+        required: true,
+    }
+
+    const descriptionValidation: Validation = {
+        required: true,
+        minLength: 10,
+        maxLength: 200,
+    }
+
+    const peopleValidation: Validation = {
+        required: true,
+        min: 1,
+        max: 5,
+    }
+
+    export class ProjectInput extends Component<HTMLDivElement, HTMLFormElement> {
+        titleInputElement: HTMLInputElement;
+        descriptionInputElement: HTMLInputElement;
+        peopleInputElement: HTMLInputElement;
+
+        constructor() {
+            super({ hostId: 'app', templateId: 'project-input', elementId: 'user-input' });
+
+            this.titleInputElement = this.querySelector('#title');
+            this.descriptionInputElement = this.querySelector('#description');
+            this.peopleInputElement = this.querySelector('#people');
+
+            this.configure();
+            this.attach('afterbegin');
+        }
+
+        private gatherUserInput(): [string, string, number] {
+            const enteredTitle = this.titleInputElement.value.trim();
+            const enteredDescription = this.descriptionInputElement.value.trim();
+            const enteredPeople = Number(this.peopleInputElement.value.trim());
+
+            return [enteredTitle, enteredDescription, enteredPeople];
+        }
+
+        private clearInputs() {
+            this.titleInputElement.value = '';
+            this.descriptionInputElement.value = '';
+            this.peopleInputElement.value = '';
+        }
+
+        @Autobind
+        private submitHandler(event: Event): void {
+            event.preventDefault();
+
+            const [title, description, people] = this.gatherUserInput();
+
+            const titleValidatable: Validatable = {
+                value: title,
+                ...titleValidation,
+            }
+
+            const descriptionValidatable: Validatable = {
+                value: description,
+                ...descriptionValidation,
+            }
+
+            const peopleValidatable: Validatable = {
+                value: people,
+                ...peopleValidation,
+            }
+
+            if (validate(titleValidatable) && validate(descriptionValidatable) && validate(peopleValidatable)) {
+                const projectState = ProjectState.getInstance();
+
+                projectState.addProject(title, description, people);
+
+                this.clearInputs();
+            } else {
+                alert('Formulário inválido!')
+            }
+        }
+
+        private configure() {
+            this.element.addEventListener('submit', this.submitHandler);
+        }
+
+        renderContent(): void {}
+    }
+}
