+++ src/decorators/autobind.decorator.ts
@@ -1,0 +1,13 @@
+namespace App {
+    export function Autobind(_target: any, _methodName: string, descriptor: PropertyDescriptor) {
+        const originalMethod = descriptor.value;
+        const newDescriptor: PropertyDescriptor = {
+            configurable: true,
+            get() {
+                return originalMethod.bind(this);
+            }
+        }
+
+        return newDescriptor;
+    }
+}
